H3C/check/requirements.txt
python-docx==0.8.11
pyyaml==6.0.1
colorama==0.4.6
jieba==0.42.1
pyinstaller==6.3.0
ttkthemes==3.2.2  # 美化界面主题
Pillow==10.2.0    # 处理图片
